backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.0.1
python-jose==3.3.0
requests==2.31.0
gunicorn==21.2.0
pytest==7.4.2
black==23.9.1
flake8==6.1.0 